Expand node.js restriction
Previously the node.js restriction was specifically version `8.4.0`, which prevents the use of any other `^8.0.0` version. This might not need to be dropped all the way down to `>=4.0.0`, but as I have a lot of projects that use 7 still, I'd prefer not to lock down to the 8 major if possible.